+++ tests/test_component_typing.py
@@ -16,7 +16,7 @@ setup_test_config()
 
 @djc_test
 class TestComponentTyping:
-    def test_data_methods_input_typed(self):
+    def test_data_methods_input_typed_custom_classes(self):
         template_called = False
         js_called = False
         css_called = False
@@ -84,6 +84,74 @@ class TestComponentTyping:
         assert js_called
         assert css_called
 
+    def test_data_methods_input_typed_default_classes(self):
+        template_called = False
+        js_called = False
+        css_called = False
+
+        class ButtonFooterSlotData(TypedDict):
+            value: int
+
+        class Button(Component):
+            class Args:
+                arg1: str
+                arg2: str
+
+            class Kwargs:
+                name: str
+                age: int
+                maybe_var: Optional[int] = None
+
+            class Slots:
+                header: SlotInput
+                footer: Optional[Slot[ButtonFooterSlotData]]
+
+            def get_template_data(self, args: Args, kwargs: Kwargs, slots: Slots, context: Context):
+                nonlocal template_called
+                template_called = True
+
+                assert isinstance(args, Button.Args)
+                assert isinstance(kwargs, Button.Kwargs)
+                assert isinstance(slots, Button.Slots)
+                assert isinstance(context, Context)
+
+            def get_js_data(self, args: Args, kwargs: Kwargs, slots: Slots, context: Context):
+                nonlocal js_called
+                js_called = True
+
+                assert isinstance(args, Button.Args)
+                assert isinstance(kwargs, Button.Kwargs)
+                assert isinstance(slots, Button.Slots)
+                assert isinstance(context, Context)
+
+            def get_css_data(self, args: Args, kwargs: Kwargs, slots: Slots, context: Context):
+                nonlocal css_called
+                css_called = True
+
+                assert isinstance(args, Button.Args)
+                assert isinstance(kwargs, Button.Kwargs)
+                assert isinstance(slots, Button.Slots)
+                assert isinstance(context, Context)
+
+            template = "<button>Click me!</button>"
+
+        assert issubclass(Button.Args, tuple)
+        assert issubclass(Button.Kwargs, tuple)
+        assert issubclass(Button.Slots, tuple)
+
+        Button.render(
+            args=Button.Args(arg1="arg1", arg2="arg2"),  # type: ignore[call-arg]
+            kwargs=Button.Kwargs(name="name", age=123),  # type: ignore[call-arg]
+            slots=Button.Slots(  # type: ignore[call-arg]
+                header="HEADER",
+                footer=Slot(lambda _ctx: "FOOTER"),
+            ),
+        )
+
+        assert template_called
+        assert js_called
+        assert css_called
+
     def test_data_methods_input_not_typed_by_default(self):
         template_called = False
         js_called = False
@@ -132,7 +200,71 @@ class TestComponentTyping:
         assert js_called
         assert css_called
 
-    def test_data_methods_output_typed(self):
+    def test_data_methods_output_typed_default_classes(self):
+        template_called = False
+        js_called = False
+        css_called = False
+
+        class Button(Component):
+            class TemplateData:
+                data1: str
+                data2: int
+
+            class JsData:
+                js_data1: str
+                js_data2: int
+
+            class CssData:
+                css_data1: str
+                css_data2: int
+
+            def get_template_data(self, args, kwargs, slots, context):
+                nonlocal template_called
+                template_called = True
+
+                return {
+                    "data1": "data1",
+                    "data2": 123,
+                }
+
+            def get_js_data(self, args, kwargs, slots, context):
+                nonlocal js_called
+                js_called = True
+
+                return {
+                    "js_data1": "js_data1",
+                    "js_data2": 123,
+                }
+
+            def get_css_data(self, args, kwargs, slots, context):
+                nonlocal css_called
+                css_called = True
+
+                return {
+                    "css_data1": "css_data1",
+                    "css_data2": 123,
+                }
+
+            template = "<button>Click me!</button>"
+
+        assert issubclass(Button.TemplateData, tuple)
+        assert issubclass(Button.JsData, tuple)
+        assert issubclass(Button.CssData, tuple)
+
+        Button.render(
+            args=["arg1", "arg2"],
+            kwargs={"name": "name", "age": 123},
+            slots={
+                "header": "HEADER",
+                "footer": Slot(lambda _ctx: "FOOTER"),
+            },
+        )
+
+        assert template_called
+        assert js_called
+        assert css_called
+
+    def test_data_methods_output_typed_custom_classes(self):
         template_called = False
         js_called = False
         css_called = False
@@ -489,43 +621,57 @@ class TestComponentTyping:
             )
 
     def test_custom_args_class_raises_on_invalid(self):
+        class Parent:
+            pass
+
         class Button(Component):
             template = "Hello"
 
-            class Args:
+            class Args(Parent):
                 arg1: str
                 arg2: str
+
+        assert issubclass(Button.Args, Parent)
+        assert not issubclass(Button.Args, tuple)
 
         with pytest.raises(TypeError, match=re.escape("Args() takes no arguments")):
             Button.render(
                 args=Button.Args(arg1="arg1", arg2="arg2"),  # type: ignore[call-arg]
             )
 
+        class Parent2:
+            def __init__(self, arg1: str, arg2: str):
+                self.arg1 = arg1
+                self.arg2 = arg2
+
         class Button2(Component):
             template = "Hello"
 
-            class Args:
+            class Args(Parent2):
                 arg1: str
                 arg2: str
 
-                def __init__(self, arg1: str, arg2: str):
-                    self.arg1 = arg1
-                    self.arg2 = arg2
+        assert not issubclass(Button2.Args, tuple)
+        assert issubclass(Button2.Args, Parent2)
 
         with pytest.raises(TypeError, match=re.escape("'Args' object is not iterable")):
             Button2.render(
                 args=Button2.Args(arg1="arg1", arg2="arg2"),
             )
 
+        class Parent3:
+            def __iter__(self):
+                return iter([self.arg1, self.arg2])  # type: ignore[attr-defined]
+
         class Button3(Component):
             template = "Hello"
 
-            class Args:
+            class Args(Parent3):
                 arg1: str
                 arg2: str
 
-                def __iter__(self):
-                    return iter([self.arg1, self.arg2])
+        assert not issubclass(Button3.Args, tuple)
+        assert issubclass(Button3.Args, Parent3)
 
         with pytest.raises(TypeError, match=re.escape("Args() takes no arguments")):
             Button3.render(
@@ -533,62 +679,77 @@ class TestComponentTyping:
             )
 
     def test_custom_args_class_custom(self):
+        class Parent:
+            def __init__(self, arg1: str, arg2: str):
+                self.arg1 = arg1
+                self.arg2 = arg2
+
+            def __iter__(self):
+                return iter([self.arg1, self.arg2])
+
         class Button(Component):
             template = "Hello"
 
-            class Args:
+            class Args(Parent):
                 arg1: str
                 arg2: str
-
-                def __init__(self, arg1: str, arg2: str):
-                    self.arg1 = arg1
-                    self.arg2 = arg2
-
-                def __iter__(self):
-                    return iter([self.arg1, self.arg2])
 
         Button.render(
             args=Button.Args(arg1="arg1", arg2="arg2"),
         )
 
     def test_custom_kwargs_class_raises_on_invalid(self):
+        class Parent:
+            pass
+
         class Button(Component):
             template = "Hello"
 
-            class Kwargs:
+            class Kwargs(Parent):
                 arg1: str
                 arg2: str
+
+        assert not issubclass(Button.Kwargs, tuple)
+        assert issubclass(Button.Kwargs, Parent)
 
         with pytest.raises(TypeError, match=re.escape("Kwargs() takes no arguments")):
             Button.render(
                 kwargs=Button.Kwargs(arg1="arg1", arg2="arg2"),  # type: ignore[call-arg]
             )
 
+        class Parent2:
+            def __init__(self, arg1: str, arg2: str):
+                self.arg1 = arg1
+                self.arg2 = arg2
+
         class Button2(Component):
             template = "Hello"
 
-            class Kwargs:
+            class Kwargs(Parent2):
                 arg1: str
                 arg2: str
 
-                def __init__(self, arg1: str, arg2: str):
-                    self.arg1 = arg1
-                    self.arg2 = arg2
+        assert not issubclass(Button2.Kwargs, tuple)
+        assert issubclass(Button2.Kwargs, Parent2)
 
         with pytest.raises(TypeError, match=re.escape("'Kwargs' object is not iterable")):
             Button2.render(
                 kwargs=Button2.Kwargs(arg1="arg1", arg2="arg2"),
             )
 
+        class Parent3:
+            def _asdict(self):
+                return {"arg1": self.arg1, "arg2": self.arg2}  # type: ignore[attr-defined]
+
         class Button3(Component):
             template = "Hello"
 
-            class Kwargs:
+            class Kwargs(Parent3):
                 arg1: str
                 arg2: str
 
-                def _asdict(self):
-                    return {"arg1": self.arg1, "arg2": self.arg2}
+        assert not issubclass(Button3.Kwargs, tuple)
+        assert issubclass(Button3.Kwargs, Parent3)
 
         with pytest.raises(TypeError, match=re.escape("Kwargs() takes no arguments")):
             Button3.render(
@@ -596,19 +757,23 @@ class TestComponentTyping:
             )
 
     def test_custom_kwargs_class_custom(self):
+        class Parent:
+            def __init__(self, arg1: str, arg2: str):
+                self.arg1 = arg1
+                self.arg2 = arg2
+
+            def _asdict(self):
+                return {"arg1": self.arg1, "arg2": self.arg2}
+
         class Button(Component):
             template = "Hello"
 
-            class Kwargs:
+            class Kwargs(Parent):
                 arg1: str
                 arg2: str
 
-                def __init__(self, arg1: str, arg2: str):
-                    self.arg1 = arg1
-                    self.arg2 = arg2
-
-                def _asdict(self):
-                    return {"arg1": self.arg1, "arg2": self.arg2}
+        assert not issubclass(Button.Kwargs, tuple)
+        assert issubclass(Button.Kwargs, Parent)
 
         Button.render(
             kwargs=Button.Kwargs(arg1="arg1", arg2="arg2"),
@@ -618,14 +783,14 @@ class TestComponentTyping:
         class Button(Component):
             template = "Hello"
 
-            class Args(NamedTuple):
+            class Args:
                 size: int
 
-            class Kwargs(NamedTuple):
+            class Kwargs:
                 color: str
 
         class ButtonExtra(Button):
-            class Args(NamedTuple):
+            class Args:
                 name: str
                 size: int
 
@@ -635,6 +800,6 @@ class TestComponentTyping:
                 assert ButtonExtra.Kwargs is Button.Kwargs
 
         ButtonExtra.render(
-            args=ButtonExtra.Args(name="John", size=30),
-            kwargs=ButtonExtra.Kwargs(color="red"),
+            args=ButtonExtra.Args(name="John", size=30),  # type: ignore[call-arg]
+            kwargs=ButtonExtra.Kwargs(color="red"),  # type: ignore[call-arg]
         )